fixes #19889 (Ошибка в работе таска SyncIntimeCommand)
git-svn-id: http://nitra.ath.cx:8080/svn/delivery_sync/trunk@66 7d45cafa-a0e2-42dc-b39b-1e02c51c2cf4

+++ src/Nitra/DeliveryBundle/Command/SyncIntimeCommand.php
@@ -25,6 +25,7 @@ class SyncIntimeCommand extends ContainerAwareCommand
         ini_set("soap.wsdl_cache_enabled", 0);
         ini_set("soap.wsdl_cache_enabled", 0);
         ini_set("soap.wsdl_cache_limit", 0);
+        ini_set('memory_limit', '1024M');
 
         header('Expires: Mon, 26 Jul 1997 05:00:00 GMT');
         header('Cache-Control: no-store, no-cache, must-revalidate');
@@ -135,8 +136,11 @@ class SyncIntimeCommand extends ContainerAwareCommand
                                      'service' => $it,
                                      'wareId' => $id
             ));
-            $department = $query->getResult()->getFirst();
-            $em->remove($department);
+			$result = $query->getResult();
+			if(count($result)) {
+				$department = $result->getFirst();
+				$em->remove($department);
+			}
         }
         $em->flush();
     }
